+++ deploy_commander/command/git.py
@@ -18,10 +18,11 @@ from fabric.context_managers import show
 from fabfile import utils
 
 
-def clone(params):
-    install_repo(params)
-
 def install_repo(params):
+    print(red("install_repo is deprecated from version 1.0.0"))
+    clone(params)
+
+def clone(params):
     """
     The clone command can be used to clone a new repo.
     If it's allready an existing path it will prompt for overwrite
@@ -42,24 +43,38 @@ def install_repo(params):
             abort("Aborted...")
         
     utils.ensure_path(params['repo_path'])
+    
     run('git clone --recursive %s %s' % (params['repo_url'], params['repo_path']))
          
     print(green("Repo `%s` successfully cloned" % params['repo_url']))
 
 def deploy_tag(params):
     """
+    Deprecated command, use deploy instead
+    """
+    
+    print(red("Warning git.deploy_tag is deprecated from version 1.0.0 !!! Use git.deploy instead!"))
+    
+    deploy(params);
+
+def deploy(params):
+    """
     This command will update a repo and copy the contents to another folder.
     It can be used to create versioned deployments of the codebase
     When executed it will prompt for the tag to deploy if it's not known
     """
-    params['repo_path'] = params['repo_path'] % env.params
+    
+    params = utils.format_params(params)
+    
+    if 'tag_path' in params:
+        params['target_path'] = params['tag_path']
+        print(red("Warning tag_path is deprecated from version 1.0.0 !!! Use target_path instead!"))
     
     if 'tag' in params:
-        env['params']['tag'] = params['tag']
-    elif 'tag' in env['params']:
-        # TODO : Check tag
-        env['params']['tag'] = env['params']['tag']
-    else:
+        params['branch'] = params['tag']
+        print(red("Warning tag is deprecated from version 1.0.0 !!! Use branch instead!"))
+        
+    if 'branch' not in params:
         with cd(params['repo_path']):
             print("Fetching latest tags...just a moment...")
             print("")
@@ -76,33 +91,29 @@ def deploy_tag(params):
                 
             print("")
         
-        env['params']['tag'] = prompt("Enter tag to deploy : ", default=latest_tag)
-    
-    params = utils.format_params(params)
-    
-    tag_path = params['tag_path']
+        params['branch'] = prompt("Enter branch/tag to deploy : ", default=latest_tag)
     
     if not exists(params['repo_path']):
         abort(red("Repo path not existing... is the project installed?"))
     
     # If exist remove full source
-    if exists(tag_path):
-        print(yellow("Deploy source path `%s` allready existed... source data be removed and reset." % tag_path))
-        run('rm -Rf %s' % tag_path)
+    if exists(params['target_path']):
+        print(yellow("Deploy target path `%s` allready existed... source data be removed and reset." % params['target_path']))
+        run('rm -Rf %s' % params['target_path'])
     
-    utils.ensure_path(tag_path)
+    utils.ensure_path(params['target_path'])
     
     with cd(params['repo_path']):
         run('git fetch')
         
         # Update local repo with latest code
-        run('git checkout %s' % (env['params']['tag']))
+        run('git checkout %s' % (params['branch']))
         
-        run('git pull origin %s' % (env['params']['tag']))
+        run('git pull origin %s' % (params['branch']))
         
         run('git submodule update')
         
         # Copy source code to version
-        run('cp -Rf %s/* %s/' % (params['repo_path'], tag_path))
+        run('cp -Rf %s/* %s/' % (params['repo_path'], params['target_path']))